chore: move SuperRefineRules into lib/types/zod
