+++ packages/toolkit/src/lib/type/index.ts
@@ -1,2 +1,3 @@
 export * from "./general";
 export * from "./apiRoute";
+export * from "./zod";
