+++ packages/toolkit/src/lib/type/zod.ts
@@ -1,0 +1,17 @@
+export type SuperRefineRule = {
+  key: string;
+  validator: (value: any) => SuperRefineRuleValidatorReturn;
+};
+
+export type SuperRefineRuleValidatorReturn =
+  | SuperRefineRuleValidatorValidReturn
+  | SuperRefineRuleValidatorInvalidReturn;
+
+export type SuperRefineRuleValidatorValidReturn = {
+  valid: true;
+};
+
+export type SuperRefineRuleValidatorInvalidReturn = {
+  valid: false;
+  error: string;
+};
