+++ packages/toolkit/src/lib/use-instill-form/type.ts
@@ -95,21 +95,3 @@ export type InstillFormTree =
   | InstillFormGroupItem
   | InstillFormConditionItem
   | InstillArrayItem;
-
-// export type SuperRefineRule = {
-//   key: string;
-//   validator: (value: any) => SuperRefineRuleValidatorReturn;
-// };
-
-// export type SuperRefineRuleValidatorReturn =
-//   | SuperRefineRuleValidatorValidReturn
-//   | SuperRefineRuleValidatorInvalidReturn;
-
-// export type SuperRefineRuleValidatorValidReturn = {
-//   valid: true;
-// };
-
-// export type SuperRefineRuleValidatorInvalidReturn = {
-//   valid: false;
-//   error: string;
-// };
