+++ packages/toolkit/src/view/pipeline-builder/use-node-input-fields/useStartOperatorTestModeInputForm.tsx
@@ -16,7 +16,7 @@ import { TextsField } from "./TextsField";
 import { ImagesField } from "./ImagesField";
 import { NumbersField } from "./NumbersField";
 import { AudiosField } from "./AudiosField";
-import { Nullable, StartOperatorBody } from "../../../lib";
+import { Nullable, StartOperatorBody, SuperRefineRule } from "../../../lib";
 
 export const useStartOperatorTestModeInputForm = (props: {
   nodes: Node<NodeData>[];
@@ -203,24 +203,6 @@ export function transformStartOperatorBodyToFields(
 
   return fields;
 }
-
-export type SuperRefineRule = {
-  key: string;
-  validator: (value: any) => SuperRefineRuleValidatorReturn;
-};
-
-export type SuperRefineRuleValidatorReturn =
-  | SuperRefineRuleValidatorValidReturn
-  | SuperRefineRuleValidatorInvalidReturn;
-
-export type SuperRefineRuleValidatorValidReturn = {
-  valid: true;
-};
-
-export type SuperRefineRuleValidatorInvalidReturn = {
-  valid: false;
-  error: string;
-};
 
 export function transformStartOperatorBodyToSuperRefineRules(
   body: Nullable<StartOperatorBody>
